Remove order_right and gate signal from arbiter
Using the 'for' loop to assign the 'token' bus introduces a precedent of
importance to the bit of the 'order' bus. The bits in the 'order' bus
that get looped over later by the 'for' loop have a higher precedence
then the earlier bits. Thus assignments due to higher precedent bits
will overwrite the assignments of lower precedent bits. No longer do we
need the gating bus to have only one bit high like what was needed in
the logical 'OR' mux.