Add intermediate signal when making 'token_lookahead'
By assigning token_lookahead directly from the the token replicate there
will always be an over assignment on the RHS. The LSB will get assigned
to the token_lookahead signal and the unused upper bits will be
discarded. This is valid verilog and the desired behaviour, however,
many tools will throw a warning because the width of the LHS and RHS do
not match.

Using an intermediate 'token_wrap' signal allows for only a slice of a
signal to be assigned to the token_lookahead, thus removing the width
discrepancy and warning.