Remove Verilator warning for variable width miss-match
An integer is a 32 bit signed number so when used with the 'ff1' for loop we
get a variable size miss-match warning from verilator. To resolve this warning
we define the 'i' variable to be the same width as the return 'ff1' variable.
However, the iterator within a "count down" for loop must be signed if the last
valid value is to be zero. Thus, we change the for loop to be "count up" to
allow the iterator to be unsigned, like the 'select' port list value.